fix(wallet-cli): route session reset/view output through CommandOutput interface
